Updated Testimonial.
Updated the css bases on changes made to the feature.

Updated the User Permission to include webmaster as being able to do everything with this feature.

Updated the view to fix the Testimonial showing up on every view pane and updated the block that goes on the home page title to be a H3 instead of a H2.

Signed-off-by: Staris <Staris@69008.no-reply.drupal.org>